api/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
flask-CORS
marshmallow-sqlalchemy
black
